backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.0.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
stripe==2.65.0
tzdata==2021.5
urllib3==1.26.8
